requirements.txt
azure-ai-language-conversations
azure-ai-vision-imageanalysis
opencv-python
